# src/main/resources/static/openapi/codylab-talent-2025-openapi.yml
openapi: 3.1.0
info:
  title: CodyLab Talent 2025 API
  description: API for managing the CodyLab Talent 2025 project.
  version: 1.0.0
servers:
  - url: http://localhost:8090/api/v1
    description: Local development server
...
paths:
...
  /users:
    get:
      summary: Retrieve all users
      description: Returns a list of User objects.
      operationId: getUsers
      tags:
        - Users
      parameters:
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/Size'
        - $ref: '#/components/parameters/Sort'
        - name: ids
          in: query
          required: false
          style: form
          explode: true
          description: List of user IDs to filter.
          schema:
            type: array
            items:
              type: integer
              format: int64
        - name: taskId
          in: query
          required: false
          description: ID of the task associated with the user.
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of users successfully retrieved.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Users'
        '204':
          description: No users found.
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
    post:
      summary: Create a new user
      description: Creates a new user and returns the created User object.
      operationId: createUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Users'
      responses:
        '201':
          description: User successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        '400':
          description: Invalid request.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /users/{id}:
    get:
      summary: Retrieve a user by their ID
      description: Returns the User object corresponding to the provided ID.
      operationId: getUserById
      tags:
        - Users
      parameters:
        - name: id
          in: path
          description: Unique identifier of the user.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User successfully retrieved.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        '404':
          description: User not found.
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
    put:
      summary: Update an existing user
      description: Updates the details of an existing user and returns the updated User object.
      operationId: updateUser
      tags:
        - Users
      parameters:
        - name: id
          in: path
          description: Unique identifier of the user to update.
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Users'
      responses:
        '200':
          description: User successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        '400':
          description: Invalid request.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
        '404':
          description: User not found.
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
    delete:
      summary: Delete a user by their ID
      description: Deletes the user corresponding to the provided ID.
      operationId: deleteUser
      tags:
        - Users
      parameters:
        - name: id
          in: path
          description: Unique identifier of the user to delete.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User successfully deleted.
        '404':
          description: User not found.
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /users/{id}/managedProjects:
    get:
      summary: Retrieve a user with their managed projects
      description: Returns the user and their managed projects
      operationId: getUserWithManagedProjects
      tags:
        - Users
      parameters:
        - name: id
          in: path
          description: User ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User with managed projects successfully retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        '404':
          description: User not found
        '500':
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
...
  parameters:
    PageNumber:
      name: pageNumber
      in: query
      description: Page number to retrieve.
      required: false
      schema:
        type: integer
        format: int32
        default: 0
        examples:
          - value: 0
            summary: Initial page
    Size:
      name: size
      in: query
      description: Number of items per page.
      required: false
      schema:
        type: integer
        format: int32
        default: 10
        examples:
          - value: 10
            summary: Page size
    Sort:
      name: sort
      in: query
      description: Sorting criteria for results.
      required: false
      schema:
        type: string
        examples:
          - value: 'id,asc'
            summary: Sort by ID in ascending order
  schemas:
...
    ApiError:
      type: object
      properties:
        details:
          type: string
          description: Detailed description of the error.
          example: Internal error
        status:
          type: integer
          format: int32
          example: 500
        error:
          type: string
          example: Internal Server Error
        message:
          type: string
        timestamp:
          type: string
          format: date-time
          example: 2025-07-02T14:00:00Z
    Projects:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the project.
        code:
          type: string
          maxLength: 10
          description: Project identification code.
        name:
          type: string
          description: Project name.
        description:
          type: string
          description: Project description.
        startDate:
          type: string
          format: date
          description: Project start date.
          examples:
            - value: '2025-01-01'
              summary: Project start date
        duration:
          type: integer
          description: Project duration in days.
        manager:
          type: string
          description: Project manager name.
...
    Users:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the user.
        firstName:
          type: string
          description: User's first name.
        lastName:
          type: string
          description: User's last name.
        username:
          type: string
          description: Username for access.
        email:
          type: string
          format: email
          description: User's email address.
        profile:
          type: string
          description: User profile (e.g., administrator, standard user).
        dailyHours:
          type: integer
          description: User's daily working hours.
        managedProjects:
          type: array
          items:
            $ref: '#/components/schemas/Projects'
...